refactor: add Moves component
